experiments/init/requirements.txt
boto3
numpy
pandas
pillow